% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_music_band.R
\name{add_music_band}
\alias{add_music_band}
\alias{ui_add_music_band}
\alias{serv_add_music_band}
\title{UI and server for adding tab}
\usage{
ui_add_music_band(modId)

serv_add_music_band(modId)
}
\arguments{
\item{modId}{id to used in the module}
}
\value{
\code{ui_add_music_band} returns ui for the tab, \code{serv_add_music_band} returns moduleServer for it
}
\description{
UI and server for adding tab
}
\examples{

}
