% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppPackage-package.R
\docType{package}
\name{AppPackage-package}
\alias{AppPackage}
\alias{AppPackage-package}
\title{AppPackage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Francesco85P/AppPackage}
  \item Report bugs at \url{https://github.com/Francesco85P/AppPackage/issues}
}

}
\author{
\strong{Maintainer}: Francesco Paparazzo \email{francescopaparazzo85@gmail.com} (\href{https://orcid.org/0009-0004-3724-1411}{ORCID})

}
\keyword{internal}
