% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Title}
\usage{
check_class(x, a)
}
\arguments{
\item{x}{object to check}

\item{a}{expected class}
}
\value{

}
\description{
Title
}
\examples{
check_class(1, "numeric")
}
