% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excavator_merge.R
\name{Excavator_merge_bed}
\alias{Excavator_merge_bed}
\title{Title}
\usage{
Excavator_merge_bed(
  file_path_target_bed,
  file_path_HSLM,
  merged_file_name = "merged_data.txt"
)
}
\arguments{
\item{file_path_target_bed}{}

\item{file_path_HSLM}{}

\item{merged_file_name}{}
}
\value{
the merged file (HSLM input for Plot_visualization)
}
\description{
Title
}
\examples{
Excavator_merge_bed(my_file_path_target_bed, my_file_path_HSLM, ,my_merged_file_name = "merged_data.txt")
}
