% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar.R
\name{find_similar}
\alias{find_similar}
\alias{ui_find_similar}
\alias{server_find_similar}
\title{Title}
\usage{
ui_find_similar(modId)

server_find_similar(modId, dbCon)
}
\arguments{
\item{modId}{id to used in the module}

\item{dbCon}{connection to database}
}
\value{

}
\description{
Title
}
\examples{

}
