% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_dict.R
\name{get_from_dict}
\alias{get_from_dict}
\title{getting information from a dict table in db}
\usage{
get_from_dict(x)
}
\arguments{
\item{x}{name of dict}
}
\value{
named vector of ids
}
\description{
getting information from a dict table in db
}
