% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_picker_input.R
\name{inline_picker_input}
\alias{inline_picker_input}
\title{Inline pickerIinput}
\usage{
inline_picker_input(id, choices, liveSearch = FALSE)
}
\arguments{
\item{id}{id of a input}

\item{choices}{choices for input}

\item{liveSearch}{use liveSearch? (default: TRUE)}
}
\value{
picketInput as inline input
}
\description{
Returns pickerInput that could be put inline as a aprt fo a sentence
}
\examples{
tags$div(
    tags$span("Someon called"),
    inline_picker_input(id = "this", choices = 11:14, liveSearch = TRUE)
)
}
