% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_id.R
\name{name_id}
\alias{name_id}
\title{Title}
\usage{
name_id(i, dbCon)
}
\arguments{
\item{i}{id}

\item{dbCon}{connection to database}
}
\value{
string containing basic information about id
}
\description{
Title
}
