% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_visualization.R
\name{Plot_Visualization}
\alias{Plot_Visualization}
\title{Plotting CNVs detection results}
\usage{
Plot_Visualization(...)
}
\arguments{
\item{...}{No input is needed to run the Shiny App}
}
\value{
A shiny App
}
\description{
Plotting CNVs detection results
}
\examples{
Plot_Visualization()
}
