% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_music_band.R
\name{add_music_band}
\alias{add_music_band}
\alias{ui_add_music_band}
\alias{serv_add_music_band}
\title{Title}
\usage{
ui_add_music_band(modId)

serv_add_music_band(modId, dbCon)
}
\arguments{
\item{modId}{id to used in the module}

\item{dbCon}{connection to database}
}
\value{

}
\description{
Title
}
\examples{

}
