% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rels.R
\name{add_rels}
\alias{add_rels}
\alias{ui_add_rels}
\alias{serv_add_rels}
\title{Title}
\usage{
ui_add_rels(modId)

serv_add_rels(modId)
}
\arguments{
\item{modId}{id to used in the module}
}
\value{

}
\description{
Title
}
\examples{

}
