% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rels.R
\name{add_rels}
\alias{add_rels}
\alias{ui_add_rels}
\alias{serv_add_rels}
\title{UI and server for adding relationships tab}
\usage{
ui_add_rels(modId)

serv_add_rels(modId)
}
\arguments{
\item{modId}{id to used in the module}
}
\value{
\code{ui_add_rels} returns ui for the tab, \code{serv_add_rels} returns moduleServer for it
}
\description{
UI and server for adding relationships tab
}
\examples{

}
