% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{checking class}
\usage{
check_class(x, a)
}
\arguments{
\item{x}{object to check}

\item{a}{expected class}
}
\value{
throws an error if object's class does not match needed. returns invisible NULL otherwise
}
\description{
checking class
}
\examples{
check_class(1, "numeric")
}
