% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar.R
\name{find_similar}
\alias{find_similar}
\alias{ui_find_similar}
\alias{server_find_similar}
\title{UI and server fo finding related/similar musicians/bands}
\usage{
ui_find_similar(modId)

server_find_similar(modId)
}
\arguments{
\item{modId}{id to used in the module}
}
\value{
\code{ui_find_similar} returns ui for the tab, \code{server_find_similar} returns moduleServer for it
}
\description{
UI and server fo finding related/similar musicians/bands
}
\examples{

}
