% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_id.R
\name{name_id}
\alias{name_id}
\title{Title}
\usage{
name_id(i, addSpecialClass = F, info = T)
}
\arguments{
\item{i}{id}

\item{addSpecialClass}{add class and present as a span?}

\item{info}{add info as in genre and type?}
}
\value{
string containing basic information about id
}
\description{
Title
}
