% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_visualization.R
\name{Plot_Visualization}
\alias{Plot_Visualization}
\title{Plot CNVs}
\usage{
Plot_Visualization(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shinyApp()}. Currently unused.}
}
\value{
Runs a Shiny application in the default web browser.
}
\description{
Plot CNVs
}
\examples{
if (interactive()) {
  Plot_Visualization()
}
}
