% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_query.R
\name{res_query}
\alias{res_query}
\title{Title}
\usage{
res_query(query, db = "mydatabase.sqlite")
}
\arguments{
\item{query}{query to execute}

\item{db}{name of ifle with database}
}
\value{

}
\description{
Title
}
